requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
httpx==0.27.2
python-dotenv==1.0.1
openai-whisper==20231117
python-multipart==0.0.9
